feat: Implement conditional TopBar and BottomBar
Refactor `MorenoFootBallNavHost` to conditionally display the `TopBar` and `BottomBar` based on the current navigation route.

The top and bottom bars will now be hidden on the `LoginScreen` and visible on other screens. This is managed by observing the `currentBackStackEntry` and checking the route name.

A new `AppScreen` composable has been introduced to encapsulate the `Scaffold` and its conditional `topBar` and `bottomBar` logic.